requirements.txt
pandas>=2.1.0,<3.0.0
streamlit>=1.30.0,<2.0.0
pdfplumber>=0.10.0,<1.0.0
google-api-python-client>=2.100.0,<3.0.0
google-auth-oauthlib>=1.1.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
plotly>=5.17.0,<6.0.0

# Dev / Test
pytest>=7.4.0,<9.0.0
